+++ game/src/main/java/net/demilich/metastone/game/spells/aura/AttributeAura.java
@@ -38,5 +38,6 @@ public final class AttributeAura extends Aura {
 				(EntityFilter) desc.get(AuraArg.FILTER),
 				desc.getCondition());
 		setDesc(desc);
+		includeExtraTriggers(desc);
 	}
 }