+++ game/src/test/java/com/hiddenswitch/spellsource/CustomCardsTests.java
@@ -53,6 +53,28 @@ import static org.testng.Assert.*;
 public class CustomCardsTests extends TestBase {
 
 	@Test
+	public void testThunderfury() {
+		runGym((context, player, opponent) -> {
+			playCard(context, player, "weapon_thunderfury");
+			player.getHero().getWeapon().setMaxHp(4);
+			player.getHero().getWeapon().setHp(4);
+			attack(context, player, player.getHero(), opponent.getHero());
+			assertEquals(player.getHero().getWeapon().getDurability(), player.getHero().getWeapon().getMaxDurability() - 1);
+			attack(context, player, player.getHero(), opponent.getHero());
+			assertEquals(player.getHero().getWeapon().getDurability(), player.getHero().getWeapon().getMaxDurability() - 2);
+			playCard(context, player, "spell_test_overload");
+			attack(context, player, player.getHero(), opponent.getHero());
+			assertEquals(player.getHero().getWeapon().getDurability(), player.getHero().getWeapon().getMaxDurability() - 2);
+			context.endTurn();
+			context.endTurn();
+			assertEquals(player.getAttributeValue(Attribute.OVERLOAD), 0);
+			assertFalse(player.getHero().getWeapon().hasAttribute(Attribute.AURA_IMMUNE));
+			attack(context, player, player.getHero(), opponent.getHero());
+			assertEquals(player.getHero().getWeapon().getDurability(), player.getHero().getWeapon().getMaxDurability() - 3);
+		});
+	}
+
+	@Test
 	public void testElorthaNoShadra() {
 		runGym((context, player, opponent) -> {
 			shuffleToDeck(context, player, "minion_ice_rager");